Release candidate.  Final reorg of classes.  moved all remaining partModule derived classes to Modules.  moved remaining root classes to proper folder locations.
