+++ BDArmory/BDArmory.csproj
@@ -72,22 +72,22 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="Control\AIUtils.cs" />
-    <Compile Include="Control\BDGenericAIBase.cs" />
-    <Compile Include="Control\BDModuleSurfaceAI.cs" />
+    <Compile Include="Modules\BDGenericAIBase.cs" />
+    <Compile Include="Modules\BDModuleSurfaceAI.cs" />
     <Compile Include="Control\IBDAIControl.cs" />
-    <Compile Include="Bullets\AmmoSwitch.cs" />
+    <Compile Include="Modules\AmmoSwitch.cs" />
     <Compile Include="FX\DecalEmitterScript.cs" />
     <Compile Include="FX\DecalGaplessParticleEmitter.cs" />
     <Compile Include="FX\MissileFX.cs" />
     <Compile Include="Guidances\CruiseGuidance.cs" />
-    <Compile Include="LoadedVesselSwitcher.cs" />
+    <Compile Include="UI\LoadedVesselSwitcher.cs" />
     <Compile Include="Misc\BahaTurretBullet.cs" />
     <Compile Include="Misc\BDAModuleInfos.cs" />
-    <Compile Include="Misc\CFEnable.cs" />
-    <Compile Include="Misc\EngageableWeapon.cs" />
-    <Compile Include="Misc\IEngageService.cs" />
+    <Compile Include="Modules\CFEnable.cs" />
+    <Compile Include="Modules\EngageableWeapon.cs" />
+    <Compile Include="Modules\IEngageService.cs" />
     <Compile Include="Modules\BDExplosivePart.cs" />
-    <Compile Include="Parts\MissileBase.cs" />
+    <Compile Include="Modules\MissileBase.cs" />
     <Compile Include="Modules\ModuleWWC.cs" />
     <Compile Include="Modules\SM_Reactive_Armor.cs" />
     <Compile Include="Modules\ModuleSpaceRadar.cs" />
@@ -95,31 +95,31 @@
     <Compile Include="UI\BDArmorySetup.cs" />
     <Compile Include="UI\BDAWindowSettingsField.cs" />
     <Compile Include="Bullets\BulletInfo.cs" />
-    <Compile Include="Parts\MissileLauncher.cs" />
+    <Compile Include="Modules\MissileLauncher.cs" />
     <Compile Include="Misc\Misc.cs" />
     <Compile Include="FX\ShellCasing.cs" />
     <Compile Include="FX\ExplosionFX.cs" />
     <Compile Include="FX\BulletHitFX.cs" />
     <Compile Include="CounterMeasure\CMFlare.cs" />
     <Compile Include="CounterMeasure\CMDropper.cs" />
-    <Compile Include="RocketLauncher.cs" />
+    <Compile Include="Modules\RocketLauncher.cs" />
     <Compile Include="FX\ParticleTurbulence.cs" />
     <Compile Include="Modules\ClusterBomb.cs" />
     <Compile Include="Misc\KSPForceApplier.cs" />
-    <Compile Include="Parts\BDModularGuidance.cs" />
+    <Compile Include="Modules\BDModularGuidance.cs" />
     <Compile Include="Modules\BDMMLauncher.cs" />
     <Compile Include="Modules\BDAdjustableRail.cs" />
     <Compile Include="FX\BDAGaplessParticleEmitter.cs" />
-    <Compile Include="MissileGuidance.cs" />
+    <Compile Include="Guidances\MissileGuidance.cs" />
     <Compile Include="Modules\Animation\BDAScaleByDistance.cs" />
     <Compile Include="Modules\Animation\BDALookConstraintUp.cs" />
     <Compile Include="Parts\SeismicChargeFX.cs" />
     <Compile Include="UI\BDGUIComboBox.cs" />
     <Compile Include="UI\BDGUIUtils.cs" />
     <Compile Include="UI\BDInputUtils.cs" />
-    <Compile Include="TargetInfo.cs" />
+    <Compile Include="Targeting\TargetInfo.cs" />
     <Compile Include="UI\BDATargetManager.cs" />
-    <Compile Include="Control\BDModulePilotAI.cs" />
+    <Compile Include="Modules\BDModulePilotAI.cs" />
     <Compile Include="Parts\TargetingCamera.cs" />
     <Compile Include="Modules\ModuleTargetingCamera.cs" />
     <Compile Include="Modules\ModuleTurret.cs" />
@@ -134,12 +134,12 @@
     <Compile Include="Parts\TGPCameraEffects.cs" />
     <Compile Include="Parts\TGPCamRotator.cs" />
     <Compile Include="CounterMeasure\CMSmoke.cs" />
-    <Compile Include="TargetSignatureData.cs" />
+    <Compile Include="Targeting\TargetSignatureData.cs" />
     <Compile Include="Modules\ModuleRadar.cs" />
     <Compile Include="Radar\RadarUtils.cs" />
     <Compile Include="Misc\BDAEditorCategory.cs" />
     <Compile Include="Misc\BDAExtensions.cs" />
-    <Compile Include="Misc\BDACategoryModule.cs" />
+    <Compile Include="Modules\BDACategoryModule.cs" />
     <Compile Include="Modules\RadarWarningReceiver.cs" />
     <Compile Include="CounterMeasure\CMChaff.cs" />
     <Compile Include="CounterMeasure\VesselChaffInfo.cs" />
@@ -152,9 +152,9 @@
     <Compile Include="UI\BDKeyBinder.cs" />
     <Compile Include="UI\BDInputSettingsFields.cs" />
     <Compile Include="UI\BDInputInfo.cs" />
-    <Compile Include="Control\ModuleWingCommander.cs" />
+    <Compile Include="Modules\ModuleWingCommander.cs" />
     <Compile Include="Misc\BDATooltips.cs" />
-    <Compile Include="MissileLaunchParams.cs" />
+    <Compile Include="Misc\MissileLaunchParams.cs" />
     <Compile Include="Misc\WMTurretGroup.cs" />
     <Compile Include="Control\BDAirspeedControl.cs" />
     <Compile Include="Misc\RippleOption.cs" />
