--- BDArmory/Bullets/AmmoSwitch.cs
@@ -1,12 +1,0 @@
-﻿using BDArmory.Core;
-using UnityEngine;
-
-namespace BDArmory.Bullets
-{
-    public class AmmoSwitch : PartModule
-    {
-
-
-
-    }
-}
