+++ BDArmory/Guidances/CruiseGuidance.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using BDArmory.Core.Extension;
 using BDArmory.Misc;
+using BDArmory.Modules;
 using BDArmory.Parts;
 using UnityEngine;
