+++ BDArmory/Guidances/MissileGuidance.cs
@@ -2,10 +2,10 @@ using System;
 using BDArmory.Core;
 using BDArmory.Core.Extension;
 using BDArmory.Misc;
-using BDArmory.Parts;
+using BDArmory.Modules;
 using UnityEngine;
 
-namespace BDArmory
+namespace BDArmory.Guidances
 {
     public class MissileGuidance
     {
