+++ BDArmory/Modules/AmmoSwitch.cs
@@ -1,0 +1,9 @@
+﻿namespace BDArmory.Modules
+{
+    public class AmmoSwitch : PartModule
+    {
+
+
+
+    }
+}
