+++ BDArmory/Modules/BDModularGuidance.cs
@@ -4,14 +4,13 @@ using BDArmory.Core;
 using BDArmory.Core.Extension;
 using BDArmory.Guidances;
 using BDArmory.Misc;
-using BDArmory.Modules;
 using BDArmory.Radar;
 using BDArmory.UI;
 using KSP.UI.Screens;
 using UniLinq;
 using UnityEngine;
 
-namespace BDArmory.Parts
+namespace BDArmory.Modules
 {
     public class BDModularGuidance : MissileBase
     {
