+++ BDArmory/Modules/RadarWarningReceiver.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using BDArmory.Core;
 using BDArmory.Misc;
 using BDArmory.Radar;
+using BDArmory.Targeting;
 using BDArmory.UI;
 using UnityEngine;
