+++ BDArmory/Radar/RadarUtils.cs
@@ -8,6 +8,7 @@ using BDArmory.UI;
 using System.Collections.Generic;
 using BDArmory.Core;
 using BDArmory.Modules;
+using BDArmory.Targeting;
 using UnityEngine;
 
 namespace BDArmory.Radar
