+++ BDArmory/Targeting/TargetInfo.cs
@@ -2,11 +2,10 @@ using System.Collections;
 using System.Collections.Generic;
 using BDArmory.Core.Extension;
 using BDArmory.Modules;
-using BDArmory.Parts;
 using BDArmory.UI;
 using UnityEngine;
 
-namespace BDArmory
+namespace BDArmory.Targeting
 {
 	public class TargetInfo : MonoBehaviour
 	{
