+++ BDArmory/UI/BDATargetManager.cs
@@ -11,6 +11,7 @@ using System.Text;
 using System;
 using BDArmory.Core;
 using BDArmory.Modules;
+using BDArmory.Targeting;
 
 namespace BDArmory.UI
 {
